Removed obsolete function, added stop button features, naming convetion fixes
We first thought, since we were having camera display problems, that something was wrong with the code and somehow we needed to specify the device used. It turns out it was just a drivers issue on both Florian's and my machine. I therefore deleted the getDevices function Florian made in an attempt to solve the issue. Also I corrected my naming of variables for coherence with Florian's naming. Finally I added some functionnality that allows the user to reuse the scanner directly instead of refreshing the page. I also removed the frozen video window that remained after the barcode was stopped.